daemon: move collection `jsonrpc_` methods to an individual module
